requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==3.0.5
channels-redis==4.1.0
charset-normalizer==3.3.0
constantly==15.1.0
cryptography==41.0.4
daphne==3.0.2
Django==4.2.6
django-chartjs==2.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ismember==1.0.2
joblib==1.3.2
msgpack==1.0.7
numpy==1.26.1
pandas==2.1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
zope.interface==6.1
